Rename photoscan file for consistency (#180)
